backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
torch>=2.0.0
pymupdf>=1.24.0
httpx>=0.27.0
python-multipart>=0.0.9
pyyaml>=6.0.1
nbformat>=5.10.0
